Initial import.
This is a test applet designed to implement the Calendar control for the Groop application.
The main goal of this project is to sucessfully connect to the Google Calendar API with authentication.

Then the plan is to interface the services available in the Google Calendar API, allowing full modification
of authenticated users Calendars.

Desired Features:
- Loading Calendars
- Creating Calendars
- Modifying Calendar Settings
- Creating Events to Calendars
- Modifying Events
- Sharing Calendars

Once this standalone applet is tested and complete, it will be integrated into the Groop application.